RSL-25: refactor code
