+++ src/components/UserStateindIcator/UserStateIndicator.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-
 import {
   Box,
   Button,
