+++ src/constants/common.js
@@ -2,3 +2,5 @@ export const DELTA = 0.00000001;
 
 export const MSEC_PER_DAY = 86400000;
 export const MSEC_PER_HOUR = 3600000;
+
+export const COUNT_ALL_WORDS = 3600;
